Set options in conftest.py instead of pytest.ini, install conftest.py
pytest.ini is ignored even if installed when testing the installed
package with 'pytest --pyargs xbout'. To get around this, set all
options in conftest.py, and move conftest.py to the xbout/ directory so
that it gets installed with the package.

Note, putting conftest.py in the xbout/ directory so that it is used
both by tests/ and calc/tests/.
