+++ setup.cfg
@@ -50,9 +50,6 @@ calc =
     xhistogram
 docs = sphinx >= 1.4
 
-[options.package_data]
-xbout = tests/pytest.ini
-
 [build_sphinx]
 project = $metadata.name
 version = $metadata.version
