--- xbout/tests/pytest.ini
@@ -1,10 +1,0 @@
-[pytest]
-filterwarnings =
-    ignore:No geometry type found, no coordinates will be added:UserWarning
-    ignore:deallocating CachingFileManager.*, but file is not already closed. This may indicate a bug\.:RuntimeWarning
-
-markers =
-    long: long test, or one of many permutations (disabled by default)
-    flaky: xarray uses this mark, adding to avoid warnings about it not being defined
-    network: xarray uses this mark, adding to avoid warnings about it not being defined
-    slow: xarray uses this mark, adding to avoid warnings about it not being defined
